Introduce Lists and other kinds
So far I have tested with workload manifests (Deployments etc.) and
multidocs consisting of those. But in general, kubeyaml needs to deal
with other kinds of resource, and with List manifests (which aggregate
resources into one document).

This runs into an apparent problem with ruamel: CommentedMap does not
necessarily produce valid YAML documents; in particular,

    d = CommentedMap()
    d.insert(0, 'foo', [], comment='same line')
    yaml.dump(d, sys.stdout)

produces the output

    foo: # same line
    []