Use a more reliable comment-adding tactic
Constructing ruamel's CommentedMap values as part of a strategy
doesn't quite work, because it appears to sometimes yield a
serialisation that ruamel cannot then parse (!). Also, it misses out
the case of whole-line comments.

Instead, add comments textually line by line. This is an advance in
that it includes whole-line comments. But, it also exposes a problem
with ruamel, so more work is needed to either avoid generating the
problematic case (until ruamel can cope with it).