Support arbitrary values in .spec.values
I wasn't checking if the field in FluxHelmResource.spec.values was a
dict before looking to see if it had an entry for 'image', so anything
other kinds of value made it explode. To make sure it works now, I've
added "noise" to the values generated when testing.

This appeared to expose more failures in ruamel, which I've tried to
work around as much as possible.