Combine top-level and "named" images in FHRs
It's completely valid to use both a top-level image and blocks
containing images in a FluxHelmRelease:

```
...
spec:
  values:
    image:
      repository: foobar/main
      tag: v1
    sidecar:
      image: foobar/side:v1
```

Thus, combine the results rather than exiting early if a top-level
image is there.

This commit also adds a check that the containers interpreted by
kubeyaml from a FluxHelmRelease resource match (in number) those
generated.

Otherwise, the tests will pass so long as `kubeyaml.containers(...)`
agrees in its interpretation with `kubeyaml.set_container(...)' --
without testing that the interpretation itself is correct.