Add test that set_fhr_image preserves structure
This test checks that setting a container's image in a custom resource
will keep the structure of the values. This is to guard against
incorrect splitting of the image name and tag, which might otherwise
not be detected by looking at the extracted containers, since doing
the extraction can reverse the incorrect split.

For example, it's possible to make the structure

    {'image': 'localhost:5000/image:tag',
     'tag': ''}

which looks equivalent to

    {'image': 'localhost:5000/image',
     'tag': 'tag'}