Add command to set the value of dot notation path
This commit adds a new `set` command, which can be used to update
one or multiple values in a resource by their dot notation path(s).

In case a given path does not resolve into a value, or the resolved
value is not a plain value, the path is collected and later returned
in an error, when all given paths have been attempted.